requirements/base.txt
# ═════════════════════════════════════════════════════════════════════════════
# Base Dependencies - Used in all environments
# ═════════════════════════════════════════════════════════════════════════════

# Django
Django==6.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1

# Database
psycopg2-binary==2.9.10

# Environment Variables
python-decouple==3.8

# CORS
django-cors-headers==4.4.0

# Filtering and Search
django-filter==24.1

# Serialization
pytz==2025.1

# Security & Encryption
PyJWT==2.10.1

# Utilities
requests==2.32.5
celery==5.4.0
redis==5.0.4

# API Documentation
drf-spectacular==0.29.0
